+++ storage-resize-images/__tests__/jest.setup.ts
@@ -1,0 +1,5 @@
+import { mockGenerateResizedImage } from "./mocks/generateResizedImage";
+
+global.config = () => require("../functions/src/config").default;
+global.deleteImage = () => require("../functions/src/config").deleteImage;
+global.mockGenerateResizedImage = mockGenerateResizedImage;