+++ storage-resize-images/__tests__/mocks/generateResizedImage.ts
@@ -1,0 +1,8 @@
+import * as functionsTestInit from "firebase-functions-test";
+
+export const mockGenerateResizedImage = () => {
+  let functionsTest = functionsTestInit();
+  return functionsTest.wrap(
+    require("../../functions/src").generateResizedImage
+  );
+};