+++ storage-resize-images/functions/src/index.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import { Bucket } from "@google-cloud/storage";
 import * as admin from "firebase-admin";
 import * as fs from "fs";
 import * as functions from "firebase-functions";
@@ -22,17 +21,15 @@ import * as mkdirp from "mkdirp";
 import * as os from "os";
 import * as path from "path";
 import * as sharp from "sharp";
-import { uuid } from "uuidv4";
 
-import config from "./config";
+import {
+  ResizedImageResult,
+  modifyImage,
+  supportedContentTypes,
+} from "./resize-image";
+import config, { deleteImage } from "./config";
 import * as logs from "./logs";
-import { ObjectMetadata } from "firebase-functions/lib/providers/storage";
-import { extractFileNameWithoutExtension } from "./util";
-
-interface ResizedImageResult {
-  size: string;
-  success: boolean;
-}
+import { extractFileNameWithoutExtension, startsWithArray } from "./util";
 
 sharp.cache(false);
 
@@ -42,24 +39,6 @@ admin.initializeApp();
 logs.init();
 
 /**
- * Supported file types
- */
-const supportedContentTypes = [
-  "image/jpeg",
-  "image/png",
-  "image/tiff",
-  "image/webp",
-];
-
-const supportedImageContentTypeMap = {
-  jpg: "image/jpeg",
-  jpeg: "image/jpeg",
-  png: "image/png",
-  tiff: "image/tiff",
-  webp: "image/webp",
-};
-
-/**
  * When an image is uploaded in the Storage bucket, we generate a resized image automatically using
  * the Sharp image converting library.
  */
@@ -67,6 +46,8 @@ export const generateResizedImage = functions.storage.object().onFinalize(
   async (object): Promise<void> => {
     logs.start();
     const { contentType } = object; // This is the image MIME type
+
+    const tmpFilePath = path.resolve("/", path.dirname(object.name)); // Absolute path to dirname
 
     if (!contentType) {
       logs.noContentType();
@@ -85,6 +66,22 @@ export const generateResizedImage = functions.storage.object().onFinalize(
 
     if (!supportedContentTypes.includes(contentType)) {
       logs.unsupportedType(supportedContentTypes, contentType);
+      return;
+    }
+
+    if (
+      config.includePathList &&
+      !startsWithArray(config.includePathList, tmpFilePath)
+    ) {
+      logs.imageOutsideOfPaths(config.includePathList, tmpFilePath);
+      return;
+    }
+
+    if (
+      config.excludePathList &&
+      startsWithArray(config.excludePathList, tmpFilePath)
+    ) {
+      logs.imageInsideOfExcludedPaths(config.excludePathList, tmpFilePath);
       return;
     }
 
@@ -144,8 +141,20 @@ export const generateResizedImage = functions.storage.object().onFinalize(
       if (failed) {
         logs.failed();
         return;
+      } else {
+        if (config.deleteOriginalFile === deleteImage.onSuccess) {
+          if (remoteFile) {
+            try {
+              logs.remoteFileDeleting(filePath);
+              await remoteFile.delete();
+              logs.remoteFileDeleted(filePath);
+            } catch (err) {
+              logs.errorDeleting(err);
+            }
+          }
+        }
+        logs.complete();
       }
-      logs.complete();
     } catch (err) {
       logs.error(err);
     } finally {
@@ -154,7 +163,7 @@ export const generateResizedImage = functions.storage.object().onFinalize(
         fs.unlinkSync(originalFile);
         logs.tempOriginalFileDeleted(filePath);
       }
-      if (config.deleteOriginalFile) {
+      if (config.deleteOriginalFile === deleteImage.always) {
         // Delete the original file
         if (remoteFile) {
           try {
@@ -169,144 +178,3 @@ export const generateResizedImage = functions.storage.object().onFinalize(
     }
   }
 );
-
-function resize(file, size) {
-  let height, width;
-  if (size.indexOf(",") !== -1) {
-    [width, height] = size.split(",");
-  } else if (size.indexOf("x") !== -1) {
-    [width, height] = size.split("x");
-  } else {
-    throw new Error("height and width are not delimited by a ',' or a 'x'");
-  }
-
-  return sharp(file)
-    .rotate()
-    .resize(parseInt(width, 10), parseInt(height, 10), {
-      fit: "inside",
-      withoutEnlargement: true,
-    })
-    .toBuffer();
-}
-
-export function convertType(buffer) {
-  const { imageType } = config;
-  if (imageType === "jpg" || imageType === "jpeg") {
-    return sharp(buffer)
-      .jpeg()
-      .toBuffer();
-  } else if (imageType === "png") {
-    return sharp(buffer)
-      .png()
-      .toBuffer();
-  } else if (imageType === "webp") {
-    return sharp(buffer)
-      .webp()
-      .toBuffer();
-  } else if (imageType === "tiff") {
-    return sharp(buffer)
-      .tiff()
-      .toBuffer();
-  }
-  return buffer;
-}
-
-const modifyImage = async ({
-  bucket,
-  originalFile,
-  fileDir,
-  fileNameWithoutExtension,
-  fileExtension,
-  contentType,
-  size,
-  objectMetadata,
-}: {
-  bucket: Bucket;
-  originalFile: string;
-  fileDir: string;
-  fileNameWithoutExtension: string;
-  fileExtension: string;
-  contentType: string;
-  size: string;
-  objectMetadata: ObjectMetadata;
-}): Promise<ResizedImageResult> => {
-  const { imageType } = config;
-  const hasImageTypeConfigSet = imageType !== "false";
-  const imageContentType = hasImageTypeConfigSet
-    ? supportedImageContentTypeMap[imageType]
-    : contentType;
-  const modifiedExtensionName =
-    fileExtension && hasImageTypeConfigSet ? `.${imageType}` : fileExtension;
-  const modifiedFileName = `${fileNameWithoutExtension}_${size}${modifiedExtensionName}`;
-  // Path where modified image will be uploaded to in Storage.
-  const modifiedFilePath = path.normalize(
-    config.resizedImagesPath
-      ? path.join(fileDir, config.resizedImagesPath, modifiedFileName)
-      : path.join(fileDir, modifiedFileName)
-  );
-  let modifiedFile: string;
-
-  try {
-    modifiedFile = path.join(os.tmpdir(), modifiedFileName);
-
-    // Cloud Storage files.
-    const metadata: { [key: string]: any } = {
-      contentDisposition: objectMetadata.contentDisposition,
-      contentEncoding: objectMetadata.contentEncoding,
-      contentLanguage: objectMetadata.contentLanguage,
-      contentType: imageContentType,
-      metadata: objectMetadata.metadata || {},
-    };
-    metadata.metadata.resizedImage = true;
-    if (config.cacheControlHeader) {
-      metadata.cacheControl = config.cacheControlHeader;
-    } else {
-      metadata.cacheControl = objectMetadata.cacheControl;
-    }
-
-    // If the original image has a download token, add a
-    // new token to the image being resized #323
-    if (metadata.metadata.firebaseStorageDownloadTokens) {
-      metadata.metadata.firebaseStorageDownloadTokens = uuid();
-    }
-
-    // Generate a resized image buffer using Sharp.
-    logs.imageResizing(modifiedFile, size);
-    let modifiedImageBuffer = await resize(originalFile, size);
-    logs.imageResized(modifiedFile);
-
-    // Generate a converted image type buffer using Sharp.
-    if (hasImageTypeConfigSet) {
-      logs.imageConverting(fileExtension, config.imageType);
-      modifiedImageBuffer = await convertType(modifiedImageBuffer);
-      logs.imageConverted(config.imageType);
-    }
-
-    // Generate a image file using Sharp.
-    await sharp(modifiedImageBuffer).toFile(modifiedFile);
-
-    // Uploading the modified image.
-    logs.imageUploading(modifiedFilePath);
-    await bucket.upload(modifiedFile, {
-      destination: modifiedFilePath,
-      metadata,
-    });
-    logs.imageUploaded(modifiedFilePath);
-
-    return { size, success: true };
-  } catch (err) {
-    logs.error(err);
-    return { size, success: false };
-  } finally {
-    try {
-      // Make sure the local resized file is cleaned up to free up disk space.
-      if (modifiedFile) {
-        logs.tempResizedFileDeleting(modifiedFilePath);
-        fs.unlinkSync(modifiedFile);
-        logs.tempResizedFileDeleted(modifiedFilePath);
-      }
-    } catch (err) {
-      logs.errorDeleting(err);
-    }
-  }
-};